+++ philanthropy/models/__init__.py
@@ -10,9 +10,11 @@ from ._wallet import ShareOfWalletRegressor
 from ._moves import MovesManagementClassifier
 from ._lapse import LapsePredictor
 from ._planned_giving import PlannedGivingIntentScorer
+from ._forecast import FinancialForecastModel
 
 __all__ = [
     "DonorPropensityModel",
+    "FinancialForecastModel",
     "LapsePredictor",
     "MajorGiftClassifier",
     "MovesManagementClassifier",